Change default array type

+++ src/gl-matrix/common.js
@@ -25,7 +25,7 @@ THE SOFTWARE. */
 
 // Configuration Constants
 export const EPSILON = 0.000001;
-export let ARRAY_TYPE = (typeof Float32Array !== 'undefined') ? Float32Array : Array;
+export let ARRAY_TYPE = (typeof Float64Array !== 'undefined') ? Float64Array : Array;
 export const RANDOM = Math.random;
 
 /**
